Added admin blueprint.
